moved the gpu functions from provider_impl into a separate rnd file
